refactor: enhance wage handling by introducing wage utility functions and updating type definitions
