+++ app/src/utils/index.ts
@@ -3,6 +3,7 @@ export * from './generalUtil'
 export * from './constantUtil'
 export * from './currencyUtil'
 export * from './expenseUtil'
+export * from './wageUtil'
 export * from './contractManagementUtil'
 export * from './contractDeploymentUtil'
 export * from './web3Util'
